+++ assets/routes/tutorial.js
@@ -1,2 +1,2 @@
-export const features = Object.freeze(["tutorial"]);
+export const features = Object.freeze(["preferences", "tutorial"]);
 export const styles = Object.freeze([]);